requirements.txt
pandas==2.2.3
yfinance==0.2.61
python-dotenv==1.1.0
matplotlib==3.10.3